requirements.txt
playsound
pygame
pygobject==3.50.0
python-dateutil
synthesizer
